fixed tests by splitting out functions into separate files to reduce circular references
