+++ src/utilities/index.ts
@@ -2,9 +2,13 @@ export * from './array'
 export * from './createMaybeRelativeUrl'
 export * from './createRouteMethods'
 export * from './createRouter'
+export * from './createRouterPush'
+export * from './flattenParentMatches'
 export * from './getInitialUrl'
+export * from './guards'
 export * from './isBrowser'
 export * from './mergeParams'
+export * from './normalizeRouteParams'
 export * from './paramValidation'
 export * from './params'
 export * from './paramsFinder'
